requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# ViT support (optional — only needed for vit_* architectures)
timm>=0.9.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.4

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Config loading
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Testing
pytest>=7.4.0
